Removed unnecessary UniqueId regeneration
While developing this code, it was initially assumed that libFuzzer would use multiple threads to run several jobs, so proper ID incrementing and Mutexes were used to ensure that each ID would be unique. Since jobs are actually generated as processes, this turned out to be an unecessary process, so it was removed.